Refactor file browser plugin to have token and interface defined in implementation instead of in plugin.
