+++ src/filebrowser/index.ts
@@ -3,3 +3,4 @@
 
 export * from './browser';
 export * from './model';
+export * from './tracker';
